% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit_sci_data.R
\docType{data}
\name{cit_sci_data}
\alias{cit_sci_data}
\title{Example citizen science data}
\format{A data frame with 10000 rows and 5 variables:
\describe{
  \item{recorder}{the ID of the person who made the observation}
  \item{species}{The ID of the species that was observed}
  \item{date}{The date the observation was made, to day precision, in date format}
  \item{long}{The longitude of the observation}
  \item{lat}{The latitude of the observation}
}}
\usage{
cit_sci_data
}
\description{
A dataset containing the 10,000 observations of citizen science data from the UK
}
\keyword{datasets}
