% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorderVisits.R
\name{recorderVisits}
\alias{recorderVisits}
\title{Calculate revisit metrics}
\usage{
recorderVisits(recorder_name, data, date_col = "dates",
  recorder_col = "recorders", location_col = "location")
}
\arguments{
\item{recorder_name}{the name of the recorder for whom you want to calculate the metrics}

\item{data}{the data.frame of recording information}

\item{date_col}{the name of the column that contains the date. This must be formatted as a date}

\item{recorder_col}{the name of the column that contains the recorder names}

\item{location_col}{the name of the column that contains the location. This is a character, such as a grid reference and should be representative of the scale at which recording is done over a single day, typically 1km-square is used.}
}
\value{
A data.frame with six columns
\itemize{
 \item{\code{recorder} - }{The name of the recorder, as given in the recorder_name argument}
 \item{\code{visits} - }{the total number of visits per recorder}
 \item{\code{sites} - }{The total number of sites visited by the recorder}
 \item{\code{PropRevisited} - }{The proportion of total sites visited that have been revisited, i.e. visited more than once, by the recorder}
 \item{\code{meanRevisits} - }{The mean number of visits per site per recorder for sites visited more than once.}
 \item{\code{SD_meanRevisits} - }{The standard deviation of the mean number of visits per site per recorder for sites visited more than once.}
}
}
\description{
This function takes in data for a recorder and calculates the revisit metrics. A visit is defined as a unique combination of recording site and date.
}
\examples{
\dontrun{

# load example data
head(cit_sci_data)

# Run for a single recorder
RV <- recorderVisits(recorder_name = 3007,
                     data = cit_sci_data,
                     date_col = 'date',
                     recorder_col = 'recorder',
                     location_col = 'location')
                   
                   
                   
# Run the metric for all recorders
RV_all <- lapply(unique(cit_sci_data$recorder),
                 FUN = recorderVisits,
                 data = cit_sci_data,
                 date_col = 'date',
                 recorder_col = 'recorder',
                 location_col = 'location')

# summarise as one table
RV_all_sum <- do.call(rbind, RV_all)

}

}
