% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_breadth.R
\name{taxa_breadth}
\alias{taxa_breadth}
\title{Metrics identifying the breadth and proportion of taxa recorded}
\usage{
taxa_breadth(recorder_name, data, sp_col = "preferred_taxon",
  recorder_col = "recorders")
}
\arguments{
\item{recorder_name}{the name of the recorder for whom you want to calculate the metrics}

\item{data}{the data.frame of recording information}

\item{sp_col}{the name of the column that contains the species names}

\item{recorder_col}{the name of the column that contains the recorder names}
}
\value{
A data.frame with four columns is returned.
\itemize{
 \item{\code{recorder} - }{The name of the recorder, as given in the recorder_name argument}
 \item{\code{taxa_breadth} - }{The total number of species recorded by this recorder}
 \item{\code{taxa_prop} - }{The proportion of species recorded by this species. Calculated as \code{taxa_breadth} divided by the total number of species recorded in \code{data}.}
 \item{\code{n} - }{The total number of observations made by this recorder}
}
}
\description{
These metrics describe the 'experience' the recorder has had recording species within the group.
}
\examples{
\dontrun{

# load example data
head(cit_sci_data)

TB <- taxa_breadth(recorder_name = 3007,
data = cit_sci_data,
sp_col = 'species',
recorder_col = 'recorder')

head(TB)

# Run for more than one recorder, this can be slow 
TB_all <- lapply(unique(cit_sci_data$recorder),
                 FUN = taxa_breadth, 
                 data = cit_sci_data, 
                 sp_col = 'species',
                 recorder_col = 'recorder')

# summarise as one table
TB_all_sum <- do.call(rbind, TB_all)

hist(TB_all_sum$taxa_prop, breaks = 40)


} 

}
